// src/app/models/table-data.model.ts
export interface TableData {
    id: number;
    brand: string;
    model: string;
    category: string;
    price: string;
    stock: number;
    rating: string;
    warranty: string;
    releaseDate: string;
    color: string;
    weight: string;
    dimensions: string;
    batteryLife: string;
    connectivity: string;
    operatingSystem: string;
    processor: string;
    gpu: string;
    description: string;
  }
  